refactor: move to helpers.ts
