+++ packages/knip/src/plugins/docusaurus/index.ts
@@ -1,7 +1,8 @@
 import type { IsPluginEnabled, Plugin, ResolveConfig } from '../../types/config.js';
 import { toDeferResolve } from '../../util/input.js';
 import { hasDependency } from '../../util/plugin.js';
-import type { ConfigItem, DocusaurusConfig, ModuleType } from './types.js';
+import { resolveConfigItem } from './helpers.js';
+import type { DocusaurusConfig } from './types.js';
 
 // https://docusaurus.io/docs/configuration
 
@@ -16,47 +17,6 @@ const config: string[] = ['docusaurus.config.{js,ts}'];
 const entry: string[] = [];
 
 const production: string[] = [];
-
-const FIRST_PARTY_MODULES = new Set([
-  'content-docs',
-  'content-blog',
-  'content-pages',
-  'debug',
-  'sitemap',
-  'svgr',
-  'rsdoctor',
-  'pwa',
-  'client-redirects',
-  'ideal-image',
-  'google-analytics',
-  'google-gtag',
-  'google-tag-manager',
-  'classic',
-  'live-codeblock',
-  'search-algolia',
-  'mermaid',
-]);
-
-const resolveModuleName = (name: string, type: ModuleType): string => {
-  if (!name.startsWith('@')) {
-    const prefix = FIRST_PARTY_MODULES.has(name) ? '@docusaurus/' : 'docusaurus-';
-    return `${prefix}${type}-${name}`;
-  }
-
-  const [scope, ...rest] = name.split('/');
-  const baseName = rest.length ? `-${rest.join('/')}` : '';
-  return `${scope}/docusaurus-${type}${baseName}`;
-};
-
-const resolveConfigItem = (item: ConfigItem, type: ModuleType): string | null => {
-  if (!item) return null;
-  const name = Array.isArray(item) ? item[0] : item;
-  if (typeof name !== 'string') return null;
-
-  // If it's already a full package name, return it
-  if (name.includes(`${type}-`)) return name;
-  return resolveModuleName(name, type);
-};
 
 const resolveConfig: ResolveConfig<DocusaurusConfig> = async config => {
   const themes = (config?.themes ?? []).map(item => resolveConfigItem(item, 'theme'));
